requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
more-itertools==10.1.0
multidict==6.0.4
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
yarl==1.9.2
